requirements.txt
anyio==3.6.2
argcomplete==1.10.3
asgiref==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.8.2
blis==0.7.9
boto3==1.26.165
botocore==1.29.165
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.3
commonmark==0.9.1
compressed-rtf==1.0.6
confection==0.0.4
constantly==15.1.0
cryptography==38.0.4
cymem==2.0.7
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
Django==3.2.19
django-cors-headers==4.0.0
django-currentuser==0.5.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
hyperlink==21.0.0
idna==3.4
IMAPClient==2.1.0
incremental==22.10.0
Jinja2==3.1.2
jmespath==1.0.1
langcodes==3.3.0
lit==16.0.5
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.3.0
murmurhash==1.0.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
packaging==23.1
pandas==2.0.1
pathy==0.10.1
pbr==5.11.1
pdfminer.six==20191110
pdfplumber==0.7.5
phonenumbers==8.12.56
pikepdf==6.2.1
Pillow==9.2.0
preshed==3.0.7
priority==1.3.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.2
Pygments==2.13.0
PyJWT==2.6.0
PyMuPDF==1.20.2
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyPDF2==2.11.1
pytesseract==0.3.10
python-crontab==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.3.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
s3transfer==0.6.1
service-identity==21.1.0
six==1.12.0
smart-open==5.2.1
sniffio==1.3.0
sockets==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-alignments==0.8.5
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
SpeechRecognition==3.8.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.3
srsly==2.4.4
starlette==0.21.0
sympy==1.12
tabulate==0.9.0
Tempita==0.5.2
termcolor==2.0.1
textract==1.6.5
thinc==8.1.2
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
tqdm==4.64.1
transformers==4.21.3
triton==2.0.0
Twisted==22.10.0
txaio==22.2.1
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
vine==5.0.0
Wand==0.6.10
wasabi==0.10.1
wcwidth==0.2.5
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==3.0.3
xlwt==0.7.2
zope.interface==5.5.0
